backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "audex-backend"
version = "0.1.0"
description = "FastAPI backend for the AUDEX MVP."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  {name = "AUDEX Team"}
]
dependencies = [
  "fastapi>=0.111,<0.112",
  "uvicorn[standard]>=0.30,<0.31",
  "pydantic>=2.7,<3.0",
  "pydantic-settings>=2.4,<3.0",
  "python-multipart>=0.0.7,<0.0.8",
  "httpx>=0.27,<0.28",
  "Pillow>=10.4,<11.0",
  "reportlab>=4.2,<5.0",
  "sqlmodel>=0.0.16,<0.0.17",
  "aiosqlite>=0.19,<0.20",
  "easyocr>=1.7,<1.8",
  "numpy>=1.26,<3.0",
  "opencv-python-headless>=4.10,<4.11",
  "ultralytics>=8.3,<9.0",
  "pdf2image>=1.17,<1.18",
  "PyMuPDF>=1.24,<1.25",
  "python-docx>=1.1,<1.2",
  "pyyaml>=6.0,<7.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9.0",
  "pytest-asyncio>=0.23,<0.24",
  "ruff>=0.4.2,<0.5"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
filterwarnings = [
  "ignore:builtin type SwigPyPacked has no __module__ attribute:DeprecationWarning",
  "ignore:builtin type SwigPyObject has no __module__ attribute:DeprecationWarning",
  "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
]
